# docs/swagger.yaml
basePath: /
definitions:
...
  dto.DashboardMonitoringResponse:
    properties:
      endpoint_id:
        type: integer
      endpoint_name:
        type: string
      monitoring_duration_days:
        type: number
      monitoring_started_at:
        type: string
    type: object
  dto.DashboardMonitoringSuccessResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.DashboardMonitoringResponse'
        type: array
      success:
        type: boolean
    type: object
  dto.DashboardOverviewResponse:
    properties:
      healthy_count:
        type: integer
      monitored_endpoints:
        type: integer
      total_endpoints:
        type: integer
      unhealthy_count:
        type: integer
    type: object
  dto.DashboardOverviewSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.DashboardOverviewResponse'
      success:
        type: boolean
    type: object
  dto.DashboardStatusResponse:
    properties:
      endpoint_id:
        type: integer
      endpoint_name:
        type: string
      monitoring_duration_days:
        type: number
      status:
        type: string
    type: object
  dto.DashboardStatusSuccessResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.DashboardStatusResponse'
        type: array
      success:
        type: boolean
    type: object
...
  dto.HistoricalReportResponse:
    properties:
      average_response_time:
        type: number
      period:
        type: string
      success_rate:
        type: number
      total_checks:
        type: integer
    type: object
  dto.HistoricalReportSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.HistoricalReportResponse'
      success:
        type: boolean
    type: object
...
  dto.IncidentResponse:
    properties:
      endpoint_id:
        type: integer
      id:
        type: integer
      is_resolved:
        type: boolean
      resolved_at:
        type: string
      started_at:
        type: string
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      access_token:
        type: string
    type: object
  dto.LoginSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.LoginResponse'
      success:
        type: boolean
    type: object
  dto.MessageResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dto.PerformanceStatsResponse:
    properties:
      average_response_time:
        type: number
      max_response_time:
        type: integer
      min_response_time:
        type: integer
    type: object
  dto.PerformanceStatsSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.PerformanceStatsResponse'
      success:
        type: boolean
    type: object
  dto.RecentIncidentsSuccessResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.IncidentResponse'
        type: array
      success:
        type: boolean
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 64
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dto.SuccessRateResponse:
    properties:
      failed:
        type: integer
      failure_rate:
        type: number
      success_rate:
        type: number
      successful:
        type: integer
      total_checks:
        type: integer
    type: object
  dto.SuccessRateSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.SuccessRateResponse'
      success:
        type: boolean
    type: object
...
  dto.UptimeReportResponse:
    properties:
      average_incident_minutes:
        type: number
      total_downtime_minutes:
        type: number
      total_incidents:
        type: integer
      uptime_percentage:
        type: number
    type: object
  dto.UptimeReportSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.UptimeReportResponse'
      success:
        type: boolean
    type: object
  dto.UserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
    type: object
  dto.UserSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.UserResponse'
      success:
        type: boolean
    type: object
...
  utils.ErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
  utils.ValidationErrorResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: Backend API for monitoring HTTP endpoints.
  title: API Performance Observatory API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT access token.
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Authenticate user
      tags:
      - Authentication
  /auth/me:
    get:
      consumes:
      - application/json
      description: Returns the currently authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account.
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ValidationErrorResponse'
      summary: Register a new user
      tags:
      - Authentication
  /dashboard/history:
    get:
      consumes:
      - application/json
      description: Returns a historical performance report.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HistoricalReportSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Historical report
      tags:
      - Dashboard
  /dashboard/incidents:
    get:
      consumes:
      - application/json
      description: Returns the most recent incidents.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RecentIncidentsSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Recent incidents
      tags:
      - Dashboard
  /dashboard/monitoring:
    get:
      consumes:
      - application/json
      description: Returns monitoring information for all endpoints.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DashboardMonitoringSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Monitoring overview
      tags:
      - Dashboard
  /dashboard/overview:
    get:
      consumes:
      - application/json
      description: Returns a summary of the monitoring dashboard.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DashboardOverviewSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Dashboard overview
      tags:
      - Dashboard
  /dashboard/performance:
    get:
      consumes:
      - application/json
      description: Returns overall response-time statistics.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PerformanceStatsSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Performance statistics
      tags:
      - Dashboard
  /dashboard/status:
    get:
      consumes:
      - application/json
      description: Returns the health status of all monitored endpoints.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DashboardStatusSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Endpoint status
      tags:
      - Dashboard
  /dashboard/success-rate:
    get:
      consumes:
      - application/json
      description: Returns overall request success statistics.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessRateSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Success rate
      tags:
      - Dashboard
  /dashboard/uptime:
    get:
      consumes:
      - application/json
      description: Returns uptime and downtime statistics.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UptimeReportSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Uptime report
      tags:
      - Dashboard
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"